finish refactoring directories and add comments to cmake files
