+++ src/core/denigma.cpp
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-#include "denigma.h"
+#include "core/denigma.h"
 #include <limits>
 
 namespace denigma {
