+++ src/core/denigma.h
