+++ src/core/finale_options.h
