+++ src/formats/enigmaxml/CMakeLists.txt
@@ -4,6 +4,8 @@ set(DENIGMA_FORMAT_ENIGMAXML_SOURCES
 )
 
 add_denigma_internal_library(denigma_format_enigmaxml ${DENIGMA_FORMAT_ENIGMAXML_SOURCES})
+# This adapter handles musx archive extraction and EnigmaXML pass-through. Keep
+# MNX/MSS/SVG/json/textmetrics dependencies out of this target.
 target_link_libraries(denigma_format_enigmaxml
     PUBLIC
         denigma_core
