+++ src/formats/mnx/mnx_converter.cpp
@@ -25,7 +25,7 @@
 #include <memory>
 #include <vector>
 
-#include "denigma.h"
+#include "core/denigma.h"
 #include "formats/enigmaxml/enigmaxml.h"
 #include "mnx.h"
