+++ src/formats/mnx/mnx_markings.h
@@ -21,7 +21,7 @@
  */
 #pragma once
 
-#include "denigma.h"
+#include "core/denigma.h"
 #include "musx/musx.h"
 #include "mnxdom.h"
