+++ src/formats/mss/mss.h
@@ -26,7 +26,7 @@
 #include <ostream>
 
 #include "denigma/conversion.h"
-#include "denigma.h"
+#include "core/denigma.h"
 
  //placeholder function
