+++ src/utils/smufl_support.h
@@ -25,7 +25,7 @@
 #include <filesystem>
 #include <optional>
 
-#include "denigma.h"
+#include "core/denigma.h"
 #include "musx/musx.h"
 
 namespace utils {
