+++ src/utils/ziputils.h
@@ -26,7 +26,7 @@
 #include <functional>
 
 #include "denigma/io/random_access_reader.h"
-#include "denigma.h"
+#include "core/denigma.h"
 
 namespace utils {
