+++ tests/mnx/test_beams.cpp
@@ -26,7 +26,7 @@
 #include <memory>
 
 #include "gtest/gtest.h"
-#include "denigma.h"
+#include "core/denigma.h"
 #include "mnxdom.h"
 #include "test_utils.h"
