+++ tests/mnx/test_scores.cpp
@@ -25,7 +25,7 @@
 #include <fstream>
 
 #include "gtest/gtest.h"
-#include "denigma.h"
+#include "core/denigma.h"
 #include "test_utils.h"
 
 using namespace denigma;
