+++ tests/mnx/test_sequences.cpp
@@ -25,7 +25,7 @@
 #include <fstream>
 
 #include "gtest/gtest.h"
-#include "denigma.h"
+#include "core/denigma.h"
 #include "formats/mnx/mnx_markings.h"
 #include "test_utils.h"
