+++ tests/test_options.cpp
@@ -23,7 +23,7 @@
 #include <filesystem>
 
 #include "gtest/gtest.h"
-#include "denigma.h"
+#include "core/denigma.h"
 #include "test_utils.h"
 
 using namespace denigma;
